requirements.txt
frida
protobuf == 3.19.3
pycryptodome
